requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
langchain-postgres==0.0.13
langchain-text-splitters==0.3.8
openai==1.78.0
psycopg[binary,pool]>=3.2.10
pgvector==0.3.6
pypdf==5.4.0
python-dotenv==1.1.0
